workflow_automation/airflow_setup.py
from airflow import DAG
from airflow.operators.python_operator import PythonOperator
from datetime import datetime

def start_ingestion():
    # Trigger NiFi ingestion
    pass

def start_processing():
    # Trigger preprocessing and feature engineering
    pass

def start_model_training():
    # Trigger model training
    pass

with DAG('customer_analytics_pipeline', start_date=datetime(2023, 1, 1), schedule_interval='@daily') as dag:
    ingestion_task = PythonOperator(task_id='ingestion', python_callable=start_ingestion)
    processing_task = PythonOperator(task_id='processing', python_callable=start_processing)
    model_training_task = PythonOperator(task_id='model_training', python_callable=start_model_training)

    ingestion_task >> processing_task >> model_training_task
