requirements.txt
pandas
tensorflow
scikit-learn
apache-airflow
requests
